--- src/main/java/student_sergejs_kibals/homework2/DogDemo.java
@@ -1,17 +1,0 @@
-package student_sergejs_kibals.homework2;
-
-public class DogDemo {
-
-    public static void main(String[] args) {
-
-        Dog Dog1 = new Dog("Killer", 5);
-
-        Dog1.voice();
-        System.out.println();
-        Dog1.voice2();
-        System.out.println();
-        Dog1.bark();
-        Dog1.happyBirthday();
-        Dog1.voice3();
-    }
-}
