+++ src/main/java/student_sergejs_kibals/homework2/level4/BankAccountDemo.java
@@ -1,4 +1,4 @@
-package student_sergejs_kibals.homework2;
+package student_sergejs_kibals.homework2.level4;
 
 class BankAccountDemo {
