+++ src/main/java/student_sergejs_kibals/homework2/level5/DogDemo.java
@@ -1,0 +1,23 @@
+package student_sergejs_kibals.homework2.level5;
+
+class DogDemo {
+
+    public static void main(String[] args) {
+
+        Dog Dog1 = new Dog("Killer", 5, "black");
+
+        Dog1.voice();
+        System.out.println();
+        Dog1.voice2();
+        System.out.println();
+        Dog1.bark();
+        Dog1.happyBirthday();
+        Dog1.voice3();
+        System.out.println();
+        Dog1.voice4();
+        System.out.println();
+        Dog1.bark();
+        Dog1.changeColor("white");
+        Dog1.voice4();
+    }
+}
